% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_check.R
\name{OM_check}
\alias{OM_check}
\title{Check input data within limits}
\usage{
OM_check(data)
}
\arguments{
\item{data}{input data. Variables must be ordered in the same way as required in the UKPDS worksheet.}
}
\value{
Warnings if any input in data exceeds limits allowed by UKPDS model
}
\description{
Check input data within limits
}
\examples{
OM_check(data)
}
