% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_clear.R
\name{OM_clear}
\alias{OM_clear}
\title{Clears UKPDS-OM2 input sheets}
\usage{
OM_clear(pathORI, pathEDIT)
}
\arguments{
\item{pathORI}{Path to UKPDS-OM2 input workbook}

\item{pathEDIT}{Path to edited workbook}
}
\value{
Cleared UKPDS-OM2 input workbook
}
\description{
Clears UKPDS-OM2 input sheets
}
\examples{
OM_clear(pathORI, pathEDIT)
}
