% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_import_outcomes.R
\name{OM_import_outcomes}
\alias{OM_import_outcomes}
\title{Imports data on outcomes from UKPDS-OM2 worksheet}
\usage{
OM_import_outcomes(path, ncore = 1)
}
\arguments{
\item{path}{Path to UKPDS-OM2 workbook}

\item{ncore}{Number of nodes in cluster to use for parallel computating}
}
\value{
Datasets containing outcomes (probability of individuals developing each endpoint in each year) over years imported from UKPDS-OM2
}
\description{
Imports data on outcomes from UKPDS-OM2 worksheet
}
\examples{
OM_import_outcomes(path, ncore = parallel::detectCores() - 1)
}
