% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_import_rf.R
\name{OM_import_rf}
\alias{OM_import_rf}
\title{Imports data on risk factors from UKPDS-OM2 worksheet}
\usage{
OM_import_rf(path, ncore = 1, baseline.only = FALSE)
}
\arguments{
\item{path}{Path to UKPDS-OM2 workbook}

\item{ncore}{Number of nodes in cluster to use for parallel computating}

\item{baseline.only}{If TRUE, imports baseline characteristics from input sheet only}
}
\value{
Datasets containing risk factors over years imported from UKPDS-OM2
}
\description{
Imports data on risk factors from UKPDS-OM2 worksheet
}
\examples{
OM_import_rf(path, ncore = parallel::detectCores() - 1, baseline.only = FALSE)
}
