% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_populate.R
\name{OM_populate}
\alias{OM_populate}
\title{Populate UKPDS-OM2 worksheets}
\usage{
OM_populate(
  pathORI,
  pathEDIT,
  data,
  sheet,
  startrow,
  startcol,
  update.input = FALSE
)
}
\arguments{
\item{pathORI}{Path to UKPDS-OM2 input workbook}

\item{pathEDIT}{Path to edited workbook}

\item{data}{Data to populate with}

\item{sheet}{Name of worksheet to populate}

\item{startrow}{Row to start population}

\item{startcol}{Column to start population}

\item{update.input}{Default set to FALSE. Set to TRUE if populating input data to update number of individuals in parameter sheet.}
}
\value{
Populated worksheet
}
\description{
Time required to populate sheet increases exponentially. Runtime for populating 1000 rows ~ 10s, 2000 rows ~ 50s, 3000 rows ~ 150s.
The recommended maximum number of patients to populate in each sheet is 3000.
}
\examples{
OM_populate(pathORI, pathEDIT, data, sheet = "Inputs", startrow = 17, startcol = 1, update.input = FALSE)
}
