% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_populate_param.R
\name{OM_populate_param}
\alias{OM_populate_param}
\title{Populate UKPDS-OM2 parameters}
\usage{
OM_populate_param(
  pathORI,
  pathEDIT,
  nprocess = 10,
  nloops = 10000,
  nboot = 0,
  nyears = 70,
  seeds = c(1:10),
  dc_qaly = 0.035,
  dc_cost = 0.035
)
}
\arguments{
\item{pathORI}{Path to UKPDS-OM2 input workbook}

\item{pathEDIT}{Path to edited workbook}

\item{nprocess}{No. of processes to use}

\item{nloops}{No. of loops (for stochastic uncertainty)}

\item{nboot}{No. of bootstraps (for parameter uncertainty)}

\item{nyears}{No. of years simulated}

\item{seeds}{Random no. seed (1 - 100)}

\item{dc_qaly}{Discount rate for QALY/LE}

\item{dc_cost}{Discount rate for costs}
}
\value{
Populated parameters sheet
}
\description{
Populate UKPDS-OM2 parameters
}
\examples{
OM_populate_param(pathORI, pathEDIT, nprocess = 10, nloops = 10000, nboot = 0, nyears = 70, seeds = c(1:10))
}
