% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_populate_rf.R
\name{OM_populate_rf}
\alias{OM_populate_rf}
\title{Populate UKPDS-OM2 risk factor sheets}
\usage{
OM_populate_rf(
  pathORI,
  pathEDIT,
  sheet,
  data = NULL,
  method = 3,
  overwrite = "N",
  tab_mc = NULL
)
}
\arguments{
\item{pathORI}{Path to UKPDS-OM2 input workbook}

\item{pathEDIT}{Path to edited workbook}

\item{sheet}{Name of risk factor worksheet}

\item{data}{Input data}

\item{method}{Method of population (1 = Match subjects initial value, 2 = y=mx + c, 3 = Use UKPDS prediction formula )}

\item{overwrite}{Replace existing values? (Y = Overwrite current values in this sheet, N = Preserve the current value of cells, populating only empty ones)}

\item{tab_mc}{Table of values for each group (m/c for continuous variables, Y/N for binary variables)}
}
\value{
Populated risk factor sheet
}
\description{
Populate UKPDS-OM2 risk factor sheets
}
\examples{
OM_populate_rf(pathORI, pathEDIT, sheet = "Smoking Status", data, method = 3, overwrite = "N", tab_mc = NULL)
}
