% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_runmacro_populate.R
\name{OM_runmacro_populate}
\alias{OM_runmacro_populate}
\title{Populate risk factors sheets}
\usage{
OM_runmacro_populate(path, check = TRUE)
}
\arguments{
\item{path}{Path to UKPDS-OM2 workbook to run macro}

\item{check}{If TRUE, check for errors and print path name to workbook if error exists}
}
\value{
Worksheets with populated risk factor trajectories
}
\description{
Populate risk factors sheets
}
\examples{
OM_runmacro_populate(path, check = TRUE)
}
