% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ukpds_colnames.R
\docType{data}
\name{ukpds_colnames}
\alias{ukpds_colnames}
\title{Column names for UKPDS-OM2}
\format{
A data frame with 29 rows and 3 variables

\itemize{
  \item \code{order}: Order of variables
  \item \code{var}: Variables used in UKPDS-OM2
  \item \code{variable}: Full description of variable (for tables, figures etc.)
}
}
\usage{
ukpds_colnames
}
\description{
Column names for UKPDS-OM2
}
\keyword{datasets}
