% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ukpds_limit.R
\docType{data}
\name{ukpds_limit}
\alias{ukpds_limit}
\title{Limits for UKPDS-OM2 inputs}
\format{
A data frame with 24 rows and 3 variables

\itemize{
  \item \code{var}: Variables used in UKPDS-OM2
  \item \code{min, max}: Minimum, maximum limit of each input within the model
}
}
\usage{
ukpds_limit
}
\description{
Limits for UKPDS-OM2 inputs
}
\keyword{datasets}
