% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ukpds_sheetnames.R
\docType{data}
\name{ukpds_sheets}
\alias{ukpds_sheets}
\title{Sheet names in UKPDS-OM2 and the corresponding variable names}
\format{
A data frame with 24 rows and 2 variables

\itemize{
  \item \code{var}: Corresponding variable names for each sheet
  \item \code{variable.sheet}: Names of sheet in UKPDS-OM2
  \item \code{var.type}: Whether variable refers to an input or an event
}
}
\usage{
ukpds_sheets
}
\description{
Sheet names in UKPDS-OM2 and the corresponding variable names
}
\keyword{datasets}
